[Finish]本の並び替え機能

+++ db/migrate/20210727093149_create_books.rb
@@ -4,7 +4,7 @@ class CreateBooks < ActiveRecord::Migration[6.1]
       t.string :title
       t.text :body
       t.integer :user_id
-      t.integer :star
+      t.float :star
       t.timestamps
     end
   end
